tests: updated E2E and unit test
