+++ test/templates/api/repositories/template-list-filter/001-404-repository/expected-http-code
@@ -1,0 +1,1 @@
+404
