+++ test/templates/api/repositories/template-list-filter/001-404-repository/expected-response.json
@@ -1,0 +1,5 @@
+{
+  "statusCode": 404,
+  "error": "templates.repositoryNotFound",
+  "message": "Repository \"unknown\" not found."
+}
