+++ test/templates/api/repositories/template-list-filter/002-404-template/expected-http-code
@@ -1,0 +1,1 @@
+404
