+++ test/templates/api/repositories/template-list-filter/002-404-template/expected-response.json
@@ -1,0 +1,5 @@
+{
+  "statusCode": 404,
+  "error": "templates.templateNotFound",
+  "message": "Template \"unknown\" not found."
+}
