+++ test/templates/api/repositories/template-list-filter/002-404-template/request.json
@@ -1,5 +1,5 @@
 {
-  "path": "/v1/repositories/keboola/templates?filter=keboola.data-apps",
+  "path": "/v1/repositories/keboola/templates/unknown/1.2.3/inputs",
   "method": "GET",
   "headers": {
     "Content-Type": "application/json",
