--- test/templates/api/repositories/template-list-filter/002-list-ok-data-apps-only/expected-http-code
@@ -1,1 +1,0 @@
-200
