--- test/templates/api/repositories/template-list-filter/002-list-ok-data-apps-only/expected-response.json
@@ -1,64 +1,0 @@
-{
-  "repository": {
-    "name": "keboola",
-    "url": "%s/test/templates/api/repositories/template-list-filter/repository",
-    "ref": "",
-    "author": {
-      "name": "Example Author",
-      "url": "https://example.com"
-    }
-  },
-  "templates": [
-    {
-      "id": "my-template-id-5",
-      "name": "My Template-snowflake-5",
-      "deprecated": false,
-      "author": {
-        "name": "Example Author",
-        "url": "https://example.com"
-      },
-      "description": "Full workflow to ...",
-      "defaultVersion": "1.2.3",
-      "categories": [
-        "Other"
-      ],
-      "components": [
-        "keboola.wr-%s",
-        "keboola.data-apps"
-      ],
-      "versions": [
-        {
-          "version": "1.2.3",
-          "stable": false,
-          "description": ""
-        }
-      ]
-    },
-    {
-      "id": "my-template-id-7",
-      "name": "My Template-snowflake-7",
-      "deprecated": false,
-      "author": {
-        "name": "Example Author",
-        "url": "https://example.com"
-      },
-      "description": "Full workflow to ...",
-      "defaultVersion": "1.2.3",
-      "categories": [
-        "Other"
-      ],
-      "components": [
-        "keboola.wr-%s",
-        "keboola.data-apps",
-        "keboola.wr-google-sheets"
-      ],
-      "versions": [
-        {
-          "version": "1.2.3",
-          "stable": false,
-          "description": ""
-        }
-      ]
-    }
-  ]
-}
