+++ test/templates/api/repositories/template-list-filter/003-404-version/expected-http-code
@@ -1,0 +1,1 @@
+404
