+++ test/templates/api/repositories/template-list-filter/003-404-version/expected-response.json
@@ -1,0 +1,5 @@
+{
+  "statusCode": 404,
+  "error": "templates.versionNotFound",
+  "message": "Version \"2.0.0\" not found."
+}
