--- test/templates/api/repositories/template-list-filter/003-list-keboola-sheet-only/expected-http-code
@@ -1,1 +1,0 @@
-200
