--- test/templates/api/repositories/template-list-filter/004-list-ok-wrong-id/expected-http-code
@@ -1,1 +1,0 @@
-200
