--- test/templates/api/repositories/template-list-filter/004-list-ok-wrong-id/expected-response.json
@@ -1,12 +1,0 @@
-{
-  "repository": {
-    "name": "keboola",
-    "url": "%s/test/templates/api/repositories/template-list-filter/repository",
-    "ref": "",
-    "author": {
-      "name": "Example Author",
-      "url": "https://example.com"
-    }
-  },
-  "templates": []
-}
