+++ test/templates/api/repositories/template-list-filter/repository/.keboola/repository.json
@@ -6,199 +6,16 @@
   },
   "templates": [
     {
-      "id": "deprecated-template-id-snowflake",
-      "name": "Deprecated Template",
-      "description": "Full workflow to ...",
-      "requirements": {
-        "backends": [
-          "snowflake"
-        ]
-      },
-      "deprecated": true,
-      "versions": [
-        {
-          "version": "1.2.3",
-          "description": "",
-          "stable": false,
-          "components": [
-            "foo.bar"
-          ]
-        }
-      ]
-    },
-    {
-      "id": "my-template-id-1",
-      "name": "My Template-teradata-1",
+      "id": "my-template-id",
+      "name": "My Template",
       "description": "Full workflow to ...",
       "path": "my-template",
-      "requirements": {
-        "backends":[
-          "teradata"
-        ]
-      },
       "versions": [
         {
           "version": "1.2.3",
           "description": "",
           "stable": false,
-          "path": "v1",
-          "components": [
-            "<keboola.wr-snowflake>",
-            "foo.bar"
-          ]
-        }
-      ]
-    },
-    {
-      "id": "my-template-id-2",
-      "name": "My Template-mysql-2",
-      "description": "Full workflow to ...",
-      "path": "my-template",
-      "requirements": {
-        "backends":[
-          "mysql"
-        ]
-      },
-      "versions": [
-        {
-          "version": "1.2.3",
-          "description": "",
-          "stable": false,
-          "path": "v1",
-          "components": [
-            "<keboola.wr-snowflake>",
-            "foo.bar"
-          ]
-        }
-      ]
-    },
-    {
-      "id": "my-template-id-3",
-      "name": "My Template-snowflake-3",
-      "description": "Full workflow to ...",
-      "path": "my-template",
-      "requirements": {
-        "backends":[
-          "snowflake"
-        ]
-      },
-      "versions": [
-        {
-          "version": "1.2.3",
-          "description": "",
-          "stable": false,
-          "path": "v1",
-          "components": [
-            "<keboola.wr-snowflake>",
-            "foo.bar"
-          ]
-        }
-      ]
-    },
-    {
-      "id": "my-template-id-4",
-      "name": "My Template-bigquery-4",
-      "description": "Full workflow to ...",
-      "path": "my-template",
-      "requirements": {
-        "backends":[
-          "bigquery"
-        ],
-        "components": [
-          "keboola.data-apps"
-        ]
-      },
-      "versions": [
-        {
-          "version": "1.2.3",
-          "description": "",
-          "stable": false,
-          "path": "v1",
-          "components": [
-            "<keboola.wr-snowflake>",
-            "foo.bar",
-            "keboola.data-apps"
-          ]
-        }
-      ]
-    },
-    {
-      "id": "my-template-id-5",
-      "name": "My Template-snowflake-5",
-      "description": "Full workflow to ...",
-      "path": "my-template",
-      "requirements": {
-        "backends":[
-          "snowflake"
-        ],
-        "components": [
-          "keboola.data-apps"
-        ]
-      },
-      "versions": [
-        {
-          "version": "1.2.3",
-          "description": "",
-          "stable": false,
-          "path": "v1",
-          "components": [
-            "<keboola.wr-snowflake>",
-            "keboola.data-apps"
-          ]
-        }
-      ]
-    },
-    {
-      "id": "my-template-id-6",
-      "name": "My Template-snowflake-6",
-      "description": "Full workflow to ...",
-      "path": "my-template",
-      "requirements": {
-        "backends":[
-          "snowflake"
-        ],
-        "components": [
-          "<keboola.wr-snowflake>"
-        ]
-      },
-      "versions": [
-        {
-          "version": "1.2.3",
-          "description": "",
-          "stable": false,
-          "path": "v1",
-          "components": [
-            "<keboola.wr-snowflake>",
-            "foo.bar"
-          ]
-        }
-      ]
-    },
-    {
-      "id": "my-template-id-7",
-      "name": "My Template-snowflake-7",
-      "description": "Full workflow to ...",
-      "path": "my-template",
-      "requirements": {
-        "backends":[
-          "snowflake"
-        ],
-        "components": [
-          "keboola.data-apps",
-          "keboola.wr-google-sheets"
-        ]
-      },
-      "versions": [
-        {
-          "version": "1.2.3",
-          "description": "",
-          "stable": false,
-          "path": "v1",
-          "components": [
-            "<keboola.wr-snowflake>",
-            "keboola.data-apps",
-            "keboola.wr-google-sheets"
-          ]
+          "path": "v1"
         }
       ]
     }
