+++ test/templates/api/repositories/template-list-filter/repository/my-template/v1/src/extractor/ex-generic-v2/shopify/config.jsonnet
@@ -1,4 +1,7 @@
 {
+  authorization: {
+    oauth_api: Input("oauth"),
+  },
   parameters: {
     token: Input("shopify-token"),
     shop: "shop",
