+++ test/templates/api/repositories/template-list-filter/repository/my-template/v1/src/extractor/keboola.ex-instagram/instagram/config.jsonnet
@@ -1,0 +1,8 @@
+{
+  authorization: {
+    oauth_api: Input("oauth2"),
+  },
+  parameters: Input("oauth2Accounts") + {
+    foo: "bar",
+  }
+}
