+++ test/templates/api/repositories/template-list-filter/repository/my-template/v1/src/extractor/keboola.ex-instagram/instagram/description.md
@@ -1,0 +1,1 @@
+test fixture
