+++ test/templates/api/repositories/template-list-filter/repository/my-template/v1/src/extractor/keboola.ex-instagram/instagram/meta.jsonnet
@@ -1,0 +1,3 @@
+{
+  name: "instagram",
+}
