+++ test/templates/api/repositories/template-list-filter/repository/my-template/v1/src/inputs.jsonnet
@@ -1,23 +1,74 @@
 {
   stepsGroups: [
     {
-      description: "Configure the eshop platforms",
+      description: "Configure your credentials for Mailchimp data source.",
       required: "all",
       steps: [
         {
-          icon: "common:settings",
-          name: "Shopify",
-          description: "Sell online with an ecommerce website",
-          inputs: [
-            {
-              id: "shopify-token",
-              name: "Shopify token",
-              description: "Please enter Shopify token",
-              type: "string",
-              kind: "hidden",
-              rules: "required",
-            },
-          ],
+          icon: "component:keboola.ex-mailchimp",
+          name: "Mailchimp",
+          description: "Mailchimp - Data Source",
+          dialogName: "Mailchimp Data Source", 
+          dialogDescription: "Extractor collects data from Mailchimp about campaigns, lists and members.",
+          inputs: []
+        }
+      ]
+    },
+    {   
+      description: "Transformations",
+      required: "all",
+      steps: [
+        {
+          icon: "component:keboola.snowflake-transformation",
+          name: "Snowflake SQL",
+          backend: "snowflake",
+          description: "Transformations",
+          inputs: [],
+        },
+        {
+          icon: "component:keboola.google-bigquery-transformation",
+          name: "BigQuery SQL",
+          backend: "bigquery",
+          description: "Transformations",
+          inputs: [],
+        },
+      ]
+    },
+    {
+      description: "Configure your credentials for data destination.",
+      required: "optional",
+      steps: [
+        {
+          icon: "component:keboola.wr-snowflake",
+          name: "Snowflake Destination",
+          description: "Load to data into snowflake",
+          dialogName: "Snowflake Destination", 
+          dialogDescription: "Data load to Snowflake DB.",
+          inputs: []
+        },
+        {
+          icon: "component:keboola.wr-google-bigquery-v2",
+          name: "Google Big Query",
+          description: "Big Query - Destination",
+          dialogName: "Big Query Destination", 
+          dialogDescription: "Data load to Google Big Query",
+          inputs: []
+        },
+        {
+          icon: "component:keboola.wr-google-sheets",
+          name: "Google sheet Destination",
+          description: "Load to data into google sheet",
+          dialogName: "Google Sheet Destination", 
+          dialogDescription: "Make a copy of [this sheet](https://docs.google.com/spreadsheets/d/14SnNVWo62MfDZPbnmuJQrg2NzKZzIzwgEg5ubRYyNLo) into your Drive. Copy text between 'spreadsheets/d/' and '/edit', paste it below",
+          inputs: [],
+        },
+        {
+          icon: "component:keboola.wr-db-pgsql",
+          name: "PostgreSQL Destination",
+          description: "Load to data into PostgreSQL",
+          dialogName: "PostgreSQL Destination", 
+          dialogDescription: "Data load to PostgreSQL DB.",
+          inputs: []
         },
       ],
     },
