+++ test/templates/api/repositories/template-use-wr-bigquery-snowflake-2/repository/snowflake/v2/src/inputs.jsonnet
@@ -8,10 +8,10 @@
             icon: 'component:keboola.wr-db-snowflake-gcs-s3',
             name: 'Snowflake',
             description: 'Destination',
+            backend: "snowflake",
             inputs: [
               {
                 id: 'wr-snowflake-host',
-                backend: "snowflake",
                 name: 'Hostname',
                 description: 'Insert database hostname',
                 type: 'string',
@@ -23,10 +23,10 @@
             icon: 'component:keboola.wr-google-bigquery-v2',
             name: 'BigQuery',
             description: 'Destination',
+            backend: "bigquery",
             inputs: [
               {
                 id: 'wr-bigquery-host',
-                backend: "bigquery",
                 name: 'Hostname',
                 description: 'Insert database hostname',
                 type: 'string',
